--- app/CommonClass/DataDictClass.py
@@ -1,46 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-    commonclass.datadict
-    ~~~~~~~~~~~~~~~~~~~~
-
-    Implements data dict
-
-    :author:    Feei <feei#feei.cn>
-    :homepage:  https://github.com/wufeifei/cobra
-    :license:   MIT, see LICENSE for more details.
-    :copyright: Copyright (c) 2017 Feei. All rights reserved
-"""
-
-__author__ = "lightless"
-__email__ = "root@lightless.me"
-
-
-class DataDict(dict):
-    def __init__(self, *args, **kwargs):
-        super(DataDict, self).__init__(*args, **kwargs)
-        for arg in args:
-            if isinstance(arg, dict):
-                for k, v in arg.iteritems():
-                    self[k] = v
-
-        if kwargs:
-            for k, v in kwargs.iteritems():
-                self[k] = v
-
-    def __getattr__(self, attr):
-        return self.get(attr)
-
-    def __setattr__(self, key, value):
-        self.__setitem__(key, value)
-
-    def __setitem__(self, key, value):
-        super(DataDict, self).__setitem__(key, value)
-        self.__dict__.update({key: value})
-
-    def __delattr__(self, item):
-        self.__delitem__(item)
-
-    def __delitem__(self, key):
-        super(DataDict, self).__delitem__(key)
-        del self.__dict__[key]
