--- app/CommonClass/__init__.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python2
-# coding: utf-8
-# file: __init__.py
-
-
-__author__ = "lightless"
-__email__ = "root@lightless.me"
-
