+++ app/controller/backend/api.py
@@ -14,9 +14,8 @@
 
 from . import ADMIN_URL
 from app import web
-from app.models import CobraRules, CobraVuls, CobraProjects
-from app.models import CobraWhiteList, CobraTaskInfo, CobraLanguages
-from app.CommonClass.ValidateClass import login_required
+from app.models import CobraRules, CobraVuls, CobraProjects, CobraWhiteList, CobraTaskInfo, CobraLanguages
+from utils.validate import login_required
 
 __author__ = "lightless"
 __email__ = "root@lightless.me"
@@ -26,7 +25,6 @@ __email__ = "root@lightless.me"
 @web.route(ADMIN_URL + '/all_rules_count', methods=['GET'])
 @login_required
 def all_rules_count():
-
     rules_count = CobraRules.query.count()
     return str(rules_count)
 
@@ -35,7 +33,6 @@ def all_rules_count():
 @web.route(ADMIN_URL + '/all_vuls_count', methods=['GET'])
 @login_required
 def all_vuls_count():
-
     vuls_count = CobraVuls.query.count()
     return str(vuls_count)
 
@@ -44,7 +41,6 @@ def all_vuls_count():
 @web.route(ADMIN_URL + '/all_projects_count', methods=['GET'])
 @login_required
 def all_projects_count():
-
     projects_count = CobraProjects.query.count()
     return str(projects_count)
 
@@ -53,7 +49,6 @@ def all_projects_count():
 @web.route(ADMIN_URL + '/all_whitelists_count', methods=['GET'])
 @login_required
 def all_whitelists_count():
-
     whitelists_count = CobraWhiteList.query.count()
     return str(whitelists_count)
 
@@ -62,7 +57,6 @@ def all_whitelists_count():
 @web.route(ADMIN_URL + '/all_tasks_count', methods=['GET'])
 @login_required
 def all_tasks_count():
-
     tasks_count = CobraTaskInfo.query.count()
     return str(tasks_count)
 
@@ -71,7 +65,5 @@ def all_tasks_count():
 @web.route(ADMIN_URL + '/all_languages_count', methods=['GET'])
 @login_required
 def all_languages_count():
-
     languages_count = CobraLanguages.query.count()
     return str(languages_count)
-
