+++ app/controller/backend/dashboard.py
@@ -19,9 +19,8 @@ from sqlalchemy.sql import func, and_
 
 from . import ADMIN_URL
 from app import web, db
-from app.CommonClass.ValidateClass import ValidateClass
-from app.models import CobraRules, CobraVuls, CobraTaskInfo
-from app.models import CobraLanguages, CobraResults, CobraProjects
+from utils.validate import ValidateClass
+from app.models import CobraRules, CobraVuls, CobraTaskInfo, CobraLanguages, CobraResults, CobraProjects
 
 __author__ = "lightless"
 __email__ = "root@lightless.me"
