+++ app/controller/backend/index.py
@@ -19,7 +19,7 @@ from flask import redirect, request, session, escape, render_template
 from sqlalchemy.sql import func, and_
 from . import ADMIN_URL
 from app import web, db
-from app.CommonClass.ValidateClass import ValidateClass, login_required
+from utils.validate import ValidateClass, login_required
 from app.models import CobraAdminUser, CobraResults, CobraProjects, CobraTaskInfo, CobraRules, CobraVuls
 from utils.common import convert_number
