+++ app/controller/backend/project.py
@@ -18,7 +18,7 @@ from flask import render_template, request, jsonify, redirect
 from . import ADMIN_URL
 from utils import config
 from app import web, db
-from app.CommonClass.ValidateClass import ValidateClass, login_required
+from utils.validate import ValidateClass, login_required
 from app.models import CobraProjects
 
 __author__ = "lightless"
