+++ app/controller/backend/vulnerability.py
@@ -17,7 +17,7 @@ from flask import request, jsonify, render_template
 
 from . import ADMIN_URL
 from app import web, db
-from app.CommonClass.ValidateClass import ValidateClass, login_required
+from utils.validate import ValidateClass, login_required
 from app.models import CobraVuls
 
 __author__ = "lightless"
