+++ app/templates/backend/project/projects.html
@@ -27,7 +27,8 @@
                     <th>Name</th>
                     <th>Author</th>
                     <th>Remark</th>
-                    <th>Last scan time</th>
+                    <th>Status</th>
+                    <th>Scan Time</th>
                     <th>Operation</th>
                 </tr>
                 </thead>
