+++ app/templates/backend/rule/rules.html
@@ -51,7 +51,7 @@
                     <th>Language</th>
                     <th>Rule type</th>
                     <th>Name</th>
-                    <th>Update Time</th>
+                    <th>Update time</th>
                     <th>Status</th>
                     <th>Operation</th>
                 </tr>
