+++ app/templates/backend/task/tasks.html
@@ -18,9 +18,9 @@
                 <tr>
                     <th>TID</th>
                     <th>Target</th>
-                    <th>Time Consume</th>
+                    <th>Consume</th>
                     <th>Status</th>
-                    <th>Last Update Time</th>
+                    <th>Update Time</th>
                     <th>Operation</th>
                 </tr>
                 </thead>
