+++ app/templates/backend/white-list/white-list.html
@@ -28,7 +28,7 @@
                     <th>RID</th>
                     <th>File</th>
                     <th>Reason</th>
-                    <th>last update time</th>
+                    <th>Update Time</th>
                     <th>Operation</th>
                 </tr>
                 </thead>
