+++ pickup/directory.py
@@ -14,12 +14,13 @@
 import time
 import os
 import logging
-from app import db, CobraExt
+from app.models import CobraExt
+from app import db
 
 logging = logging.getLogger(__name__)
 
 
-class Directory:
+class Directory(object):
     def __init__(self, path):
         self.path = path
