+++ utils/const.py
@@ -13,7 +13,7 @@
 """
 
 
-class Vulnerabilities:
+class Vulnerabilities(object):
     def __init__(self, key):
         self.key = key
